postgres/requirements.txt
psycopg==3.1.18
python-dotenv==1.0.1